==> notebooks/1_data_explorer.ipynb <==
# Exploring Data from the API

import pandas as pd
import requests as req

from settings import TOKEN

HEADER = {
    "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64; rv:78.0) Gecko/20100101 Firefox/78.0",
    "Accept-Language": "en-US;q=0.7,es-MX;q=0.3",
    "Accept-Charset": "application/x-www-form-urlencoded; charset=UTF-8",
    "Origin": "https://developer.riotgames.com",
    "X-Riot-Token": TOKEN
    }

## Get summoners per league

def get_entries(queue: str, tier: str, division: str, pages: int):
    
    base_url = f"https://euw1.api.riotgames.com/lol/league/v4/entries/{queue}/{tier}/{division}?page={pages}" 
    #RANKED_SOLO_5x5/DIAMOND/I?page=1"
    res = req.get(base_url, headers=HEADER)
    
    return res.json()

entries =  get_entries("RANKED_SOLO_5x5", "DIAMOND", "I", 1)

entries[0]

### We are interested on the following columns

- summonerId
- summonerName
- tier
- rank
- leaguePoints
- wins
- loses

**summonerId** is useful to get extra account information

## Get summoners account ids

def get_acc_info(s_id: str):
    
    base_url = f"https://euw1.api.riotgames.com/lol/summoner/v4/summoners/{s_id}" 
    res = req.get(base_url, headers=HEADER)
    
    return res.json()

acc_info = get_acc_info(entries[0]["summonerId"])

acc_info

**accountId** is useful to get the player's match history

## Get match history by summoner

- needs accountId
- needs queue id (420 for 5v5 Ranked Solo Games)
- needs begin time and end time 1593561600 first of july
- needs number of games (we multiply it by 100 to set the beginIndex)

def get_matchlist(a_id: str, begin_time = 1593561600, queue_id = 420):
    base_url = f"https://euw1.api.riotgames.com/lol/match/v4/matchlists/by-account/{a_id}?queue={queue_id}" 
    res = req.get(base_url, headers=HEADER)
    
    return res.json()

matchlist = get_matchlist("Xqb-V8M3R76oVcB07Xl9HIW1aQQtlFBdIc0SUN4QJMCvVpg")

matchlist["matches"][0]

**gameId** can be used to retrieve all the data from a specific match

## Request match data with match id

def get_match_data(m_id: str):
    base_url = f"https://euw1.api.riotgames.com/lol/match/v4/matches/{m_id}" 
    res = req.get(base_url, headers=HEADER)
    
    return res.json()

match = get_match_data(matchlist["matches"][0]["gameId"])

match.keys()

### The following keys are useful for us

- gameId
- participantIdentities
    - player
        - summonerId
- participants
    - stats
    - timeline

match["participants"][0]["stats"].keys()

match["participantIdentities"][0]["player"].keys()

### We need to drop unnecesary data in the stats columns

stats = pd.DataFrame([x["stats"] for x in match["participants"]])

stats = stats.drop(columns=[ 'item0',
 'item1',
 'item2',
 'item3',
 'item4',
 'item5','item6', "unrealKills" ,'combatPlayerScore',
 'objectivePlayerScore',
 'totalPlayerScore',
 'totalScoreRank','playerScore0',
 'playerScore1',
 'playerScore2',
 'playerScore3',
 'playerScore4',
 'playerScore5',
 'playerScore6',
 'playerScore7',
 'playerScore8',
 'playerScore9','perk0',
 'perk0Var1',
 'perk0Var2',
 'perk0Var3',
 'perk1',
 'perk1Var1',
 'perk1Var2',
 'perk1Var3',
 'perk2',
 'perk2Var1',
 'perk2Var2',
 'perk2Var3',
 'perk3',
 'perk3Var1',
 'perk3Var2',
 'perk3Var3',
 'perk4',
 'perk4Var1',
 'perk4Var2',
 'perk4Var3',
 'perk5',
 'perk5Var1',
 'perk5Var2',
 'perk5Var3',
 'perkPrimaryStyle',
 'perkSubStyle',
 'statPerk0',
 'statPerk1',
 'statPerk2'])
