==> notebooks/2_players_pool.ipynb <==
# Players Pool Handler

import time

from itertools import chain

import requests as req
import pandas as pd
import numpy as np

from settings import TOKEN

## 1. Master leagues

def get_players_master():
    container = []
    leagues = [
        "challengerleagues",
        "grandmasterleagues",
        "masterleagues"
    ]
    base_url = "https://euw1.api.riotgames.com/lol/league/v4"
    for league in leagues:  
        url = f"{base_url}/{league}/by-queue/RANKED_SOLO_5x5?api_key={TOKEN}"
        res = req.get(url)
        if res.status_code == 200:
            container.append(res.json())
        else:
            print(f"Something went wrong {res.status_code}")
            break
    return container

- Execute the function

players_master = get_players_master()

- We get an array of the top 3 leagues
- Each league has its own number of entries which are the players
- In the players object there is no "tier" attribute, we need to attacht it

print(players_master[0].keys())
players_master[0]["entries"][0]

- Attaching the tier to each entry

for x in players_master:
    for y in x["entries"]:
        y["tier"] = x["tier"]

players_master[0]["entries"][0]

- Now we want to flatten all leagues into one list because we already specified the league on each entry

players_master = list(chain.from_iterable([x["entries"] for x in players_master]))

len(players_master)

## 2. Student leagues (This time only Diamond + )

- Each page gives us around 200 players
- We should loop on a range and break the loop once the number of players retrieved is 0

def get_players_student(queue: str, tiers: dict):
    base_url = "https://euw1.api.riotgames.com/lol/league/v4/entries"
    student_players = []
    n_requests = 0
    for tier in tiers:
        for division in tiers[tier]:
            pages = 0
            while True:
                # control the number of requests
                if n_requests >= 100:
                    print(f"Number of Entries so far: {len(student_players)}")
                    n_requests = 0
                    time.sleep(121)
                pages += 1
                url = f"{base_url}/{queue}/{tier}/{division}?page={pages}&api_key={TOKEN}"
                res = req.get(url)
                n_requests += 1
                if res.status_code == 200:
                    if len(res.json()) > 0:
                        student_players.append(res.json())
                    else:
                        break
                else:
                    print(f"Something went wrong {res.status_code}")
                    break
                    
    return student_players

- Execute the function

# first one is a lot of entries.
#players_student = get_players_student("RANKED_SOLO_5x5", {"DIAMOND": ["I", "II"], "PLATINUM": "I"})
players_student = get_players_student("RANKED_SOLO_5x5", {"DIAMOND": "I"})

len(players_student)

- Now we want to flatten all the pages into one big list.

chained = list(chain.from_iterable(players_student))

chained[0]

len(chained)

## 3. Give it a clean format and merge both lists

### The important data points:

- summonerId
- summonerName
- leaguePoints
- rank
- wins
- losses
- tier

- Drop uncommon columns

players_df = pd.DataFrame(players_student).drop(columns=["leagueId", "miniSeries", "queueType"])

players_df.shape

- Concatenate the data into one big player list

players_df = pd.concat([players_df, pd.DataFrame(players_master)])

## 4. Check the data

print(f"Columns: {players_df.shape[1]} Rows: {players_df.shape[0]}\n")
print(f"Number of duplcated rows: {players_df.duplicated().sum()}\n")
print("Null values per column:\n")
print(players_df.isna().sum())
print("Data types:\n")
print(players_df.dtypes)

- Remove duplicated rows

players_df.loc[players_df["summonerId"] == "vL_isz55-r376XwG2YK8avzmHmYN815ZuwxvohHxXM3PeVTS"]

players_df = players_df.drop_duplicates()

print(f"Columns: {players_df.shape[1]} Rows: {players_df.shape[0]}\n")
print(f"Number of duplcated rows: {players_df.duplicated().sum()}\n")
print("Null values per column:\n")
print(players_df.isna().sum())

## 5. Drop useless columns

- veteran
- inactive
- freshBlood
- hotStreak

players_df = players_df.drop(columns=["veteran", "inactive", "freshBlood", "hotStreak"])

players_df.head()

## 6. Export the data

players_df.to_pickle("raw_data/players_pool.pkl", protocol=4)
