==> notebooks/3_players_account_info.ipynb <==
# Players account info handler

import time

from itertools import chain

import requests as req
import pandas as pd
import numpy as np

from settings import TOKEN

## Goal
Retrieve the account_id from the players account info becuase with it we can start retrieving the match history of each player

## 1. Import players pool data

players_pool = pd.read_pickle("raw_data/players_pool.pkl")
players_pool.head()

id_list = players_pool["summonerId"].to_list()
id_list[0]

## 2. Check how the api works

def get_acc_info(summoner_id: str):
    
    url = f"https://euw1.api.riotgames.com/lol/summoner/v4/summoners/{summoner_id}?api_key={TOKEN}"
    res = req.get(url)
    
    return res.json()

get_acc_info(id_list[0])

- We are interested on storing the accountId value into the players data

# 3. Adjust the function to retrieve each individual account information

- max 100 request every 2 minutes

def get_accounts_info(id_list: list, token: str, path: str, start=0, end=0) -> pd.DataFrame:
    container = []
    base_url = "https://euw1.api.riotgames.com/lol/summoner/v4/summoners"
    n_requests = 0
    while start < end:
        # request account info for each summoner_id
        for i in range(start, end):
            # control the number of requests
            if n_requests >= 100:
                print(f"Number of Entries so far: {len(container)}")
                n_requests = 0
                time.sleep(121)

            # make request
            url = f"{base_url}/{id_list[i]}?api_key={token}"
            res = req.get(url)
            if res.status_code == 200:
                container.append(res.json())
                # update position
                start += 1
            else:
                # if error save to csv
                pd.DataFrame(container).to_csv(
                    f"{path}/fragmented_data/account_info_{start}.pkl", index=False)
                print(f"Something went wrong {res.status_code}")
                break
            # update the number of requests
            n_requests += 1
    print(f"Done, number of elements: {len(container)}")
    # return useful data
    container = pd.DataFrame(container).rename(columns={"id": "summonerId"})
    container = container[["summonerId", "accountId"]]
    container.to_pickle(f"{path}/account_info.pkl", protocol=4)
    return f"File saved at {path}/account_info.pkl"

- Execute the function and hope for the best :D

get_accounts_info(id_list, TOKEN, "raw_data", 0, 100)

## 4. Check the data (already done in the function above this is just a step by step)

first = pd.DataFrame(first)
print(f"Columns: {first.shape[1]} Rows: {first.shape[0]}\n")
print(f"Number of duplcated rows: {first.duplicated().sum()}\n")
print("Null values per column:\n")
print(first.isna().sum())

- Rename the id column to match summonerId on the player's pool

first = first.rename(columns={"id": "summonerId"})

- Select the keys we want

first = first[["summonerId", "accountId"]]

## 5. Merge the data from accounts to player_pool data

players_pool = players_pool.merge(first, on="summonerId")

players_pool.head()

## 6. Export the data

players_pool.to_pickle("raw_data/players_pool_acc_id.pkl", protocol=4)

# As a function

def merge_with(players_pool: pd.DataFrame, container: pd.DataFrame, path: str):
    players_pool = players_pool.merge(container, on="summonerId")
    players_pool.to_pickle(f"{path}/players_pool_account.pkl", protocol=4)
    return f"File saved at {path}/players_pool_account.pkl"

acc_info = pd.read_pickle("raw_data/account_info.pkl")
acc_info.head()

merge_with(players_pool, acc_info, "raw_data")

- Check the file

merged = pd.read_pickle("raw_data/players_pool_account.pkl")
merged.head()
