==> notebooks/5_matches_data.ipynb <==
# Get full matches info using match id

import time

from itertools import chain

import requests as req
import pandas as pd
import numpy as np

from settings import TOKEN

## Goal

Add the matches found to our database collection so that we can perform analysis based on it.

## 1. Check our matches id data

matches_id = pd.read_pickle("../raw_data/matches_id.pkl")
print(f"Has {matches_id.shape[0]} unique matches")
matches_id.tail()

## 2. Make a request to the API so that we can check what kind of information we are getting as response
    
We should focus specially on the **participants** entries and clean it a bit, some columns are better for champion and matches statistics instead of player statistics, so lets split the data up.

match_id_try = matches_id["gameId"]

raw_matches = []

for id in match_id_try[:10]:
    res = req.get(f"https://euw1.api.riotgames.com/lol/match/v4/matches/{id}?api_key={TOKEN}")
    raw_matches.append(res.json())

raw_matches[0].keys()

## 3. Extract match information

We are looking for the following format 


| match_id |	region | date_created |	match_duration | patch | winner |
| --- | --- | --- | --- | --- | --- |
| 4671787510 | 	EUW1 |	2020-06-21 | 1557 |	10.12 | Red |

**Basic match info**

match_keys = ["gameId", "platformId", "gameCreation", "gameDuration", "gameVersion"]

matches_list = [{key: match[key] for key in match if key in match_keys} for match in raw_matches]
matches_list[0]

**Winners**

winners = [team["teamId"] for x in raw_matches for team in x["teams"] if team["win"] == "Win"]
winners

**Merge the lists**

matches_df = pd.DataFrame(matches_list)
matches_df = matches_df.assign(winner=winners)
matches_df["winner"] = matches_df["winner"].replace({100:"Blue", 200: "Red"})
matches_df.head()

**Clean up a bit the data to make it readable**

- Change gameCreation to DateTime format


matches_df["gameCreation"] = pd.to_datetime(matches_df["gameCreation"], unit="ms").dt.to_period("D")
matches_df.head()

**Change column names to a better format**

matches_df.rename(columns={
    "gameId": "match_id",
    "platformId": "region",
    "gameCreation": "date_created",
    "gameDuration": "match_duration",
    "gameVersion": "patch"
}, inplace=True)
matches_df.head()

**Change patch format to a double digit e.g. 10.12**

matches_df["patch"] = matches_df["patch"].str.slice(stop=5)
matches_df.head()

## 4. Extract Champions data

**Get banned champions for each match**

| champion_id | match_id | banned | 
| --- | --- | --- |

print(raw_matches[0]["gameId"])
print([x["teamId"] for x in raw_matches[0]["teams"]])
teams = [x["teamId"] for x in raw_matches[0]["teams"]]
print("\nGet the bans for each team:\n")
print({n: raw_matches[0]["teams"][n]["bans"] for n in [i for i, x in enumerate(teams)]})

bans = []
for i, x in enumerate(raw_matches):
    for y in x["teams"]:
        for z in y["bans"]:
            bans.append({"champion_id": z["championId"],
                         "match_id": x["gameId"], 
                         "banned": 1})
bans[0]

bans = pd.DataFrame(bans)
bans.head()

**Get picked champions for each match**

| champion_id | match_id | region | picked | lane | role | opponent | won | lost |
| --- | --- | --- | --- | --- | --- | --- | --- | --- |
| 62 | 4671787510 | euw1 | 1 | JUNGLE | SOLO | 43 | 1 | 0 |

print(raw_matches[0]["gameId"])
print(raw_matches[0]["participants"][0]["championId"])
print(raw_matches[0]["participants"][0]["timeline"]["lane"])

**Get the data we want**

- ids, region, lane, opponents, win or lose, rune_set, item_set.

picks = []
for x in raw_matches:
    for i, y in enumerate(x["participants"]):
        picks.append({
            "champion_id": y["championId"],
            "match_id": x["gameId"],
            "region": x["platformId"],
            "picked": 1,
            "lane": y["timeline"]["lane"],
            "role": y["timeline"]["role"],
            "opponent": ([x["championId"] 
                              for x in x["participants"][5:]] 
                                 if i < 5 
                                 else [x["championId"] 
                                    for x in x["participants"][:5]]),
            "won": y["stats"]["win"],
            "lost": not y["stats"]["win"],
        })
picks[0]

picks = pd.DataFrame(picks)
picks.head()

**Replace True, False with 1, 0 respectively**

picks = picks.assign(won=np.where(picks["won"], 1, 0), lost=np.where(picks["lost"], 1, 0))

picks.head()

**Now we need to explode the dataframe to get each individual matchup**

picks = picks.explode(column="opponent").reset_index()
picks.drop(columns="index", inplace=True)
picks.head()

**Next we need to check the lane and role variables, we want this format for role**

 - TOP
 - JUNGLE
 - MIDDLE
 - BOTTOM
 - SUPPORT

support = picks.loc[picks["role"] == "DUO_SUPPORT"].index
middle = picks.loc[picks["role"] == "DUO"].index

picks.loc[picks.index[support], 'lane'] = "SUPPORT"
picks.loc[picks.index[middle], 'lane'] = "MIDDLE"

picks = picks.drop(columns="role")
picks.head()

[x for x in picks["lane"].unique()]

## 5. Extract players data

### 5.1 Player Info

- extract players information
- extract players champion information
- extract players lane information

players_info = pd.read_pickle("../raw_data/players_pool_account.pkl")
players_info = players_info.reindex(columns=["accountId", "summonerId", "summonerName", "tier", "rank", "leaguePoints", "wins", "losses"])
players_info = players_info.rename(columns={"accountId": "account_id",
                             "summonerId": "summoner_id",
                             "summonerName": "name",
                             "leaguePoints": "league_points"})
players_info.head()

**Getting the account info, champion used and role.**

info_keys = ["accountId", "summonerId", "currentPlatformId", "summonerName"]
print([x for x in raw_matches[0]["participantIdentities"][0]["player"].items() if x[0] in info_keys])
print(raw_matches[0]["participants"][0]["championId"])
print(raw_matches[0]["participants"][0]["timeline"]["lane"])
print(raw_matches[0]["participants"][0]["timeline"]["role"])

participants = []
for x in raw_matches:
    for i, y in enumerate(x["participants"]):
        participants.append({
            "account_id": x["participantIdentities"][i]["player"]["accountId"],
            "summoner_id": x["participantIdentities"][i]["player"]["summonerId"],
            "region": x["participantIdentities"][i]["player"]["currentPlatformId"],
            "name": x["participantIdentities"][i]["player"]["summonerName"],
            "champion": y["championId"],
            "lane": y["timeline"]["lane"],
            "role": y["timeline"]["role"],
        })
participants[0]

**Create a dataframe, and assign the right values to lane variable**

participants_df = pd.DataFrame(participants)

support = participants_df.loc[participants_df["role"] == "DUO_SUPPORT"].index
middle = participants_df.loc[participants_df["role"] == "DUO"].index
participants_df.loc[participants_df.index[support], 'lane'] = "SUPPORT"
participants_df.loc[participants_df.index[middle], 'lane'] = "MIDDLE"

participants_df = participants_df.drop(columns="role")
participants_df.head()

**Now lets create frequency colums for each lane per account_id as a new table**

players_lane = participants_df[["account_id", "lane"]]
players_lane.head()

**Do the same for champions**

    Lets convert the champion ids to names, for this we need to request the champion list from the api!

champions = req.get("https://ddragon.leagueoflegends.com/cdn/10.14.1/data/en_US/champion.json").json()

champion_keys = champions["data"].keys()
champions_list = {}
for x in champions["data"]:
    champions_list[int(champions["data"][x]["key"])] = champions["data"][x]["name"]

champions_list
participants_df = participants_df.replace({"champion": champions_list})
participants_df["champion"].head()

**Creating the DataFrame**

players_champions = participants_df[["account_id", "champion"]]
players_champions.head()

**Isolate the account_info table**

    Drop the lane and champion columns and drop duplicated because we dont need them

participants_df = participants_df.drop(columns=["lane", "champion"])
participants_df.head(1)

participants_df = participants_df.drop_duplicates()
participants_df.head()

**Update the players information table**

players_info = players_info.merge(participants_df, on=["account_id", "summoner_id", "name"], how="left")

players_info.head()

### 5.2 Laning stats @10

laning = []
for x in raw_matches:
    for i, y in enumerate(x["participants"]):
        laning.append({
            "account_id": x["participantIdentities"][i]["player"]["accountId"],
            "region": x["participantIdentities"][i]["player"]["currentPlatformId"],
            "champion": y["championId"],
            "lane": y["timeline"]["lane"],
            "role": y["timeline"]["role"],
            "xppm_10": y["timeline"]["xpPerMinDeltas"]["0-10"],
            "cspm_10": y["timeline"]["creepsPerMinDeltas"]["0-10"],
            "goldpm_10": y["timeline"]["goldPerMinDeltas"]["0-10"],
            "dmg_takenpm_10": y["timeline"]["damageTakenPerMinDeltas"]["0-10"],
        })
laning[0]

players_laning_stats = pd.DataFrame(laning)

support = players_laning_stats.loc[players_laning_stats["role"] == "DUO_SUPPORT"].index
middle = players_laning_stats.loc[players_laning_stats["role"] == "DUO"].index
players_laning_stats.loc[players_laning_stats.index[support], 'lane'] = "SUPPORT"
players_laning_stats.loc[players_laning_stats.index[middle], 'lane'] = "MIDDLE"

players_laning_stats = players_laning_stats.drop(columns="role")
players_laning_stats.head()

### 5.3 Combat stats


combat = []
for x in raw_matches:
    for i, y in enumerate(x["participants"]):
        combat.append({
            "account_id": x["participantIdentities"][i]["player"]["accountId"],
            "region": x["participantIdentities"][i]["player"]["currentPlatformId"],
            "champion": y["championId"],
            "lane": y["timeline"]["lane"],
            "role": y["timeline"]["role"],
            "dmg_total": y["stats"]["totalDamageDealtToChampions"],
            "healing_total": y["stats"]["totalHeal"],
            "units_healed": y["stats"]["totalUnitsHealed"],
            "damage_mitigated": y["stats"]["damageSelfMitigated"],
            "crowd_control": y["stats"]["totalTimeCrowdControlDealt"],
            "dmg_taken": y["stats"]["totalDamageTaken"],
            "first_blood": y["stats"]["firstBloodKill"],
            "first_blood_assist": y["stats"]["firstBloodAssist"],
        })
combat[0]

player_combat_stats = pd.DataFrame(combat)

support = player_combat_stats.loc[player_combat_stats["role"] == "DUO_SUPPORT"].index
middle = player_combat_stats.loc[player_combat_stats["role"] == "DUO"].index
player_combat_stats.loc[player_combat_stats.index[support], 'lane'] = "SUPPORT"
player_combat_stats.loc[player_combat_stats.index[middle], 'lane'] = "MIDDLE"

player_combat_stats = (player_combat_stats.assign(
                        first_blood=np.where(player_combat_stats["first_blood"], 1, 0),
                        first_blood_assist=np.where(player_combat_stats["first_blood_assist"], 1, 0)))
player_combat_stats = player_combat_stats.drop(columns="role")
player_combat_stats.head()

### 5.4 Flair stats


flair = []
for x in raw_matches:
    for i, y in enumerate(x["participants"]):
        flair.append({
            "account_id": x["participantIdentities"][i]["player"]["accountId"],
            "region": x["participantIdentities"][i]["player"]["currentPlatformId"],
            "champion": y["championId"],
            "lane": y["timeline"]["lane"],
            "role": y["timeline"]["role"],
            "killing_sprees": y["stats"]["killingSprees"],
            "longest_time_alive": y["stats"]["longestTimeSpentLiving"],
            "double_kills": y["stats"]["doubleKills"],
            "triple_kills": y["stats"]["tripleKills"],
            "quadra_kills": y["stats"]["quadraKills"],
            "penta_kills": y["stats"]["pentaKills"],
        })
flair[0]

player_flair_stats = pd.DataFrame(flair)

support = player_flair_stats.loc[player_flair_stats["role"] == "DUO_SUPPORT"].index
middle = player_flair_stats.loc[player_flair_stats["role"] == "DUO"].index
player_flair_stats.loc[player_flair_stats.index[support], 'lane'] = "SUPPORT"
player_flair_stats.loc[player_flair_stats.index[middle], 'lane'] = "MIDDLE"

player_flair_stats = player_flair_stats.drop(columns="role")
player_flair_stats.head()

### 5.5 Objective stats

objectives = []
for x in raw_matches:
    for i, y in enumerate(x["participants"]):
        objectives.append({
            "account_id": x["participantIdentities"][i]["player"]["accountId"],
            "region": x["participantIdentities"][i]["player"]["currentPlatformId"],
            "champion": y["championId"],
            "lane": y["timeline"]["lane"],
            "role": y["timeline"]["role"],
            "dmg_to_objectives": y["stats"]["damageDealtToObjectives"],
            "dmg_to_turrets": y["stats"]["damageDealtToTurrets"],
            "total_cs": y["stats"]["totalMinionsKilled"],
            "jungle_cs": y["stats"]["neutralMinionsKilled"],
            "jungle_invaded": y["stats"]["neutralMinionsKilledEnemyJungle"],
            "wards_placed": y["stats"]["wardsPlaced"],
            "wards_killed": y["stats"]["wardsKilled"],
        })
objectives[0]

player_objective_stats = pd.DataFrame(objectives)

support = player_objective_stats.loc[player_objective_stats["role"] == "DUO_SUPPORT"].index
middle = player_objective_stats.loc[player_objective_stats["role"] == "DUO"].index
player_objective_stats.loc[player_objective_stats.index[support], 'lane'] = "SUPPORT"
player_objective_stats.loc[player_objective_stats.index[middle], 'lane'] = "MIDDLE"

player_objective_stats = player_objective_stats.drop(columns="role")
player_objective_stats.head()

# On the next chapter I clean all of this mess into functions
