==> notebooks/6_merging_data_files.ipynb <==
import os
import numpy as np
import pandas as pd

path = "../data/partial_files/"
elements_list = ["players_info", "match_info", "players_lanes", "player_laning_stats",
                 "player_flair_stats", "champion_bans", "champion_picks",
                 "player_combat_stats", "player_objective_stats", "players_champions"]
files_list = os.listdir(path)

# Concatenate the dataframes found for each element

frames = {}
for element in elements_list:
    to_concat = []
    for file in files_list:
        if element in file:
            to_concat.append(pd.read_pickle(path+file))
    frames[element] = pd.concat(to_concat, ignore_index=True)

# Check the data for duplicates or weird values

frames.keys()

frames["match_info"].duplicated().sum()

frames["players_info"].duplicated().sum()

frames["players_champions"].head()

frames["players_lanes"].head()

frames["player_laning_stats"].duplicated().sum()

frames["player_combat_stats"].duplicated().sum()

frames["player_objective_stats"].duplicated().sum()

frames["player_flair_stats"].duplicated().sum()

frames["champion_bans"]["champion"] = frames["champion_bans"]["champion"].replace(-1, "no ban")

frames["champion_picks"].duplicated().sum()

# Clean the data from duplicates and check dtypes and store the cleaned data to our data folder

final_path = "../data/"

**players_info**

frames["players_info"].dtypes

frames["players_info"] = frames["players_info"].drop_duplicates()
frames["players_info"].to_csv(final_path+"players_info.csv", index=False)

**match_info**

frames["match_info"].dtypes
frames["match_info"].to_csv(final_path+"match_info.csv", index=False)

**champion_statistics**

frames["champion_bans"].to_csv(final_path+"champion_bans.csv", index=False)
frames["champion_picks"].to_csv(final_path+"champion_picks.csv", index=False)

**player_statistics**

frames["players_lanes"].to_csv(final_path+"player_lanes.csv", index=False)
frames["players_champions"].to_csv(final_path+"players_champions.csv", index=False)

# in game
# drop duplicates
frames["player_laning_stats"] = frames["player_laning_stats"].drop_duplicates()
frames["player_combat_stats"] = frames["player_combat_stats"].drop_duplicates()
frames["player_objective_stats"] = frames["player_objective_stats"].drop_duplicates()
frames["player_flair_stats"] = frames["player_flair_stats"].drop_duplicates()


# store files
frames["player_laning_stats"].to_csv(final_path+"player_laning_stats.csv", index=False)
frames["player_combat_stats"].to_csv(final_path+"player_combat_stats.csv", index=False)
frames["player_objective_stats"].to_csv(final_path+"player_objective_stats.csv", index=False)
frames["player_flair_stats"].to_csv(final_path+"player_flair_stats.csv", index=False)

**Merge the data**

shared = ["match_id", "account_id", "region", "champion", "lane", "won"]

complete_df = (pd.merge(frames["player_laning_stats"], frames["player_combat_stats"], on=shared, how="left")
                .merge(frames["player_objective_stats"], on=shared, how="left")
                .merge(frames["player_flair_stats"], on=shared, how="left")
                .fillna(0))
complete_df.head()

complete_df.to_pickle("../data/merged_stats.pkl", protocol=4)
