==> notebooks/7_to_mysql_server.ipynb <==
import pandas as pd

import pymysql as sql
import settings as env

connection = sql.connect(
    host=env.HOST,
    user=env.USERNAME,
    password=env.PASSWORD,
    db=env.DB,
    cursorclass=sql.cursors.DictCursor
)

cursor = connection.cursor()

query = """
SELECT * FROM player_combat_stats
LEFT JOIN player_flair_stats
USING (account_id, champion, lane, won)
LEFT JOIN player_laning_stats
USING (account_id, champion, lane, won)
LEFT JOIN player_objective_stats
USING (account_id, champion, lane, won)
WHERE lane = 'TOP';
"""
cursor.execute(query)
data = cursor.fetchall()

pd.DataFrame(data)
