==> notebooks/7_train_ml.ipynb <==
# Detect the best variables for each role so that we have variables to compare performance between a random player and our dataset

from datetime import datetime, timedelta
from functools import reduce

import numpy as np
import pandas as pd

pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', 20)

import sklearn.linear_model as linear
import sklearn.tree as tree
import sklearn.ensemble as rf
import sklearn.svm as svm
import sklearn.neural_network as neural

import sklearn.feature_selection as feat
import sklearn.metrics as metrics

from sklearn.preprocessing import StandardScaler
from sklearn.model_selection import train_test_split
from sklearn.model_selection import KFold


# Prediction using isolated player ingame statistics

match_info = pd.read_csv("../data/match_info.csv")
laning = pd.read_csv("../data/player_laning_stats.csv")
combat = pd.read_csv("../data/player_combat_stats.csv")
flair = pd.read_csv("../data/player_flair_stats.csv")
objectives = pd.read_csv("../data/player_objective_stats.csv")

**Define a function that handles the cleaning and prediction for us:**

def get_prediction(data: pd.DataFrame, lane, model, key_features=[], train=0.9, random_seed=12, feature_selection=0):

    # selecting the lanes and dropping non useful columns
    data = data.loc[laning["lane"] == lane]
    data = data.drop(columns=["match_id","account_id", "region", "champion", "lane"])
    try:
        data = data.drop(columns=["patch", "date_created"])
    except:
        print("sorting columns not found.")
    
    # defining our target and variables
    target = data["won"]

    if len(key_features) > 0:
        variables = data[key_features]
    else:
        variables = data.loc[:, data.columns != "won"]

    # creating a list of columns so that we can return the top features
    columns = variables.columns.to_list()

    # standarazing our variables
    scale = StandardScaler()
    scale.fit(variables)
    variables = scale.transform(variables)
    del(scale)

    # splitting our test and train data
    variables_train, variables_test, target_train, target_test = train_test_split(variables, target, train_size=train, random_state=random_seed)

    # training the model
    model = model()
    model.fit(variables_train, target_train);

    # implementing feature selection if needed

    try:
        if feature_selection > 0:

            # recursive feature selection
            rfe = feat.RFE(model, n_features_to_select=feature_selection);
            rfe.fit(variables_train, target_train);
    except: 
        feature_selection = 0
    
    # returning multiple variables

    results = {
            "accuracy": round(model.score(variables_test, target_test), 3),
            #"balanced_accuracy":  round(metrics.balanced_accuracy_score(target_test, model.predict(variables_test)), 3),
            #"precision": round(metrics.precision_score(target_test, model.predict(variables_test)), 3),
            #"avg_precision": round(metrics.average_precision_score(target_train, model.predict(variables_train)), 3),
            "key_features": [columns[index] for index, ranking in enumerate(rfe.ranking_) if ranking < 4] if feature_selection > 0 else "No feature selection",
            }

    return results

## 1. Laning stats

get_prediction(laning, "TOP", tree.DecisionTreeClassifier)

## 2. Combat stats

get_prediction(combat, "TOP", tree.DecisionTreeClassifier)

## 3. Objective stats

get_prediction(objectives, "TOP", tree.DecisionTreeClassifier)

## 4. Flair stats

get_prediction(flair, "TOP", tree.DecisionTreeClassifier)

**From the above examples we see that we cannot have an accurate prediction using isolated statistics, we need more data, lets then combine all the player's ingame statistics**

# Prediction with merged player ingame statistics

## 1. Merge the stats and make a prediction for each role

shared = ["match_id", "account_id", "region", "champion", "lane", "won"]

complete_df = (pd.merge(laning, combat, on=shared, how="left")
                .merge(objectives, on=shared, how="left")
                .merge(flair, on=shared, how="left")
                .fillna(0))

for x in ["TOP", "JUNGLE", "MIDDLE", "BOTTOM", "SUPPORT"]:
    print(f"{x}: {get_prediction(complete_df, x, rf.RandomForestClassifier, random_seed=12)}")
    del(x)

## 2. Lets find the best features for TOP lane

key_features = get_prediction(complete_df, "TOP", rf.RandomForestClassifier, random_seed=12, feature_selection=9)["key_features"]

key_features

get_prediction(complete_df, "TOP", rf.RandomForestClassifier, random_seed=12, key_features=key_features)

print(f"{len(key_features)} features selected out of {len(complete_df.drop(columns=['account_id', 'region', 'champion', 'lane']).columns)}")

# Using time and patch variables to see if that increases our accuracy

match_info["patch"] = pd.to_numeric(match_info["patch"], errors="coerce")
match_info.head()

sorted_df = complete_df.merge(match_info[["match_id", "patch", "date_created"]], on="match_id", how="left").dropna()
sorted_df.head()

## 1. By Patch

last_patch = sorted_df.loc[sorted_df["patch"] == 10.14]
patches = sorted_df.loc[sorted_df["patch"] > 10.12]
patches_3 = sorted_df.loc[sorted_df["patch"] >= 10.12]

**Last patch**

for x in ["TOP", "JUNGLE", "MIDDLE", "BOTTOM", "SUPPORT"]:
    print(f"{x}: {get_prediction(last_patch, x, rf.RandomForestClassifier, random_seed=12, sorted=True)}")
    del(x)

**Last 2 patches**

for x in ["TOP", "JUNGLE", "MIDDLE", "BOTTOM", "SUPPORT"]:
    print(f"{x}: {get_prediction(patches, x, rf.RandomForestClassifier, random_seed=12, sorted=True)}")
    del(x)

**Last 3 patches**

for x in ["TOP", "JUNGLE", "MIDDLE", "BOTTOM", "SUPPORT"]:
    print(f"{x}: {get_prediction(patches_3, x, rf.RandomForestClassifier, random_seed=12, sorted=True)}")
    del(x)

## 2. By date

def get_date(days: int) -> str:
    since =  pd.to_datetime(sorted_df["date_created"].max()).date() - timedelta(days=days)
    if since.day < 10:
        day = f"0{since.day}"
    else:
        day = since.day
    if since.month < 10:
        month = f"0{since.month}"
    else:
        month = since.month

    since = f"{since.year}-{month}-{day}"
    return since

last_month = sorted_df.loc[sorted_df["date_created"] > get_date(30)]
two_weeks = sorted_df.loc[sorted_df["date_created"] > get_date(14)]
one_week = sorted_df.loc[sorted_df["date_created"] > get_date(7)]

**Last month**

for x in ["TOP", "JUNGLE", "MIDDLE", "BOTTOM", "SUPPORT"]:
    print(f"{x}: {get_prediction(last_month, x, rf.RandomForestClassifier, random_seed=12, sorted=True)}")
    del(x)

**Last two weeks**

for x in ["TOP", "JUNGLE", "MIDDLE", "BOTTOM", "SUPPORT"]:
    print(f"{x}: {get_prediction(two_weeks, x, rf.RandomForestClassifier, random_seed=12, sorted=True)}")
    del(x)

**Last week**

for x in ["TOP", "JUNGLE", "MIDDLE", "BOTTOM", "SUPPORT"]:
    print(f"{x}: {get_prediction(one_week, x, rf.RandomForestClassifier, random_seed=12, sorted=True)}")
    del(x)

# Run multiple algorithms with and without features and store the results so that we can make a better analysis

def get_model_accuracy():

    # store results on a dictionary for future analysis
    model_accuracy = {
        "model": ["rf_classifier", "linear_ridge", "linear_logistic", "linear_svc", "linear_stochastic", "decision_tree", "neural_network", "support_vc"],
        "accuracy_avg": [],
    }

    # define the models to use
    models = {
        "rf_classifier": rf.RandomForestClassifier,
        "linear_ridge": linear.RidgeClassifier,
        "linear_logistic" : linear.LogisticRegression,
        "linear_svc": svm.LinearSVC,
        "linear_stochastic": linear.SGDClassifier,
        "decision_tree": tree.DecisionTreeClassifier,
        "neural_network": neural.MLPClassifier,
        "support_vc": svm.SVC,
    }

    # define the lanes
    lanes = ["TOP", "JUNGLE", "MIDDLE", "BOTTOM", "SUPPORT"]

    # make predictions without features
    for i, model in enumerate(models):
        results = []
        # return mean avg score without features
        for lane in lanes:
            prediction = get_prediction(last_month, lane, models[model], sorted=True)
            results.append(prediction["accuracy"])

        # append mean prediction result to model_accuracy
        model_accuracy["accuracy_avg"].append(float(format(np.mean(results), ".2f")))
        print(f"Done at {i}")

    print("Done without features")

    return model_accuracy

model_accuracy = get_model_accuracy()

model_accuracy = pd.DataFrame(model_accuracy)
model_accuracy

**From the accuracy average I determined that RandomForestClassifier was the best approach, since it is not as high cost as support vector classification or Neural Networks**

model_accuracy.to_pickle("../data/model_accuracy.pkl", protocol=4)

## Feature selection and period accuracy

def get_model_acc_period():

    model_by_period = {
        "period": ["complete", "last_patch", "last_2_patches", "last_3_patches", "last_month", "last_two_weeks", "last_week"],
        "TOP": [],
        "JUNGLE": [],
        "MIDDLE": [],
        "BOTTOM": [],
        "SUPPORT": [],
    } 

    lane_features = {
        "TOP": [],
        "JUNGLE": [],
        "MIDDLE": [],
        "BOTTOM": [],
        "SUPPORT": [],
    }

    # define the iterations
    periods = {
        "complete": complete_df,
        "last_patch": last_patch,
        "last_2_patches": patches,
        "last_3_patches": patches_3,
        "last_month": last_month, 
        "last_two_weeks": two_weeks, 
        "last_week": one_week
        }

    # define the lanes
    lanes = ["TOP", "JUNGLE", "MIDDLE", "BOTTOM", "SUPPORT"]

    # without features
    for period in periods:
        for lane in lanes:
            prediction = get_prediction(periods[period], lane, rf.RandomForestClassifier)
            model_by_period[lane].append(prediction["accuracy"])

    for lane in lane_features:
        prediction = get_prediction(last_month, lane, rf.RandomForestClassifier, feature_selection=7)
        lane_features[lane].append(prediction["key_features"])

    return [model_by_period, lane_features]

results = get_model_acc_period()

results[0]

pd.DataFrame(results[0]).to_pickle("../data/model_by_period.pkl", protocol=4)

pd.DataFrame(results[1]).to_pickle("../data/lane_features.pkl", protocol=4)
