==> notebooks/9_radars.ipynb <==
import plotly.graph_objects as go
import pandas as pd
import numpy as np

with_key_values = pd.read_pickle("../data/with_key_values.pkl")

with_key_values["top_winners"].head()

# Getting the average

keys = ["Combat", "Objectives", "Laning", "Supporting", "Jungling", "Survivability"]

mid_winners = with_key_values["mid_winners"]
mid_winners_mean = mid_winners[keys].mean()

mid_winners["average"] = 0
for x in mid_winners[keys].columns:
    mid_winners["average"] = mid_winners["average"] + mid_winners[x]

mid_winners["average"] = mid_winners["average"].div(len(mid_winners[keys].columns))

mid_winners = mid_winners.sort_values(by="average", ascending=False).reset_index().drop(columns="index")

mid_winners

timmy = mid_winners[keys].iloc[400]
pro = mid_winners[keys].iloc[0]

# Comparing timmy to the average

fig = go.Figure(
  layout_title_text="Timmy vs Average",
  layout_width=1360,
  layout_height=768,
  )

fig.add_trace(go.Scatterpolar(
      r=mid_winners_mean,
      theta=keys,
      fill='toself',
      name='Average',
      marker_color="gray"
))


fig.add_trace(go.Scatterpolar(
      r=timmy,
      theta=keys,
      fill='toself',
      name='Timmy',
      marker_color="red",
))


fig.update_layout(
  polar=dict(
    radialaxis=dict(
      visible=True,
      range=[0, 18.5],
    )),
  showlegend=True,
)

fig.show()

# Comparing Timmy to a Pro

fig = go.Figure(
  layout_title_text="Timmy vs Pro",
  layout_width=1360,
  layout_height=768,
  )

fig.add_trace(go.Scatterpolar(
      r=pro,
      theta=keys,
      fill='toself',
      name='Pro',
      marker_color="gray",
))


fig.add_trace(go.Scatterpolar(
      r=timmy,
      theta=keys,
      fill='toself',
      name='Timmy',
      marker_color="red",
))


fig.update_layout(
  polar=dict(
    radialaxis=dict(
      visible=True,
      range=[0, 18.5],
    )),
  showlegend=True,
)

fig.show()

# NOW COMPARTE TIMMY KEY PERFORMANCE INDICATOR TO PRO KEY PERFORMANCE

mid_features = pd.read_pickle("../data/middle_features.pkl")
mid_features = [x for x in mid_features if x != "healing_total"]

pro = mid_winners[mid_features].iloc[0]
timmy = mid_winners[mid_features].iloc[400] 

keyson = ["Total Damage Dealt", "Damage Mitigated", "Damage avoided", "Damage to Neutral Objectives", "Damage to Turrets", "Total CS", "Enemy Jungle Invasion", "Longest Time Alive"]

fig = go.Figure(
    data=[
        go.Bar(
            name="Timmy",
            y=timmy,
            x= keyson,
            marker=dict(color='red')
            ),
        go.Bar(
            name="Pro",
            y=pro,
            x= keyson,
            marker=dict(color='gray')
            ),
        ],
    layout_title_text="Timmy vs Pro KPIs",
    layout_width=1360,
    layout_height=768
)
fig.update_yaxes(autorange=False, range=[0, 21])
fig.show()
